[Frontend] Extract PluginConfigForm and mount in detail page
Pulls all form-state, rendering, and save logic out of PluginConfigDialog
into a new reusable PluginConfigForm component. The dialog now thin-wraps
the form, and PluginDetail mounts it inline (with page-level unsaved-changes
protection) when the plugin is installed. Task 16 will remove the dialog
entirely.
